requirements.txt
google-api-python-client>=1.0.0
google-auth-httplib2>=0.0.1
google-auth-oauthlib>=0.0.1
watchdog>=0.10.0
python-dotenv>=0.15.0